migrate performance snippets
