+++ docs/website/docs/conftest.py
@@ -7,11 +7,12 @@ from dlt.common.utils import set_working_dir
 from dlt.common.configuration.container import Container
 
 # patch which providers to enable
-from dlt.common.configuration.providers import ConfigProvider, EnvironProvider, SecretsTomlProvider, ConfigTomlProvider
+from dlt.common.configuration.providers import StringTomlProvider, EnvironProvider, SecretsTomlProvider, ConfigTomlProvider
 from dlt.common.configuration.specs.config_providers_context import ConfigProvidersContext, ConfigProvidersConfiguration
 
 from tests.utils import patch_home_dir, autouse_test_storage, preserve_environ, duckdb_pipeline_location, wipe_pipeline
 
+string_toml_provider = StringTomlProvider("")
 
 @pytest.fixture(autouse=True)
 def setup_tests(request):
@@ -19,9 +20,12 @@ def setup_tests(request):
     dname =  os.path.dirname(request.module.__file__)
     config_dir = dname + "/.dlt"
 
+    # clear string toml provider
+    string_toml_provider.update("")
+
     # inject provider context so the original providers are restored at the end
     def _initial_providers():
-        return [EnvironProvider(), SecretsTomlProvider(project_dir=config_dir, add_global_config=False), ConfigTomlProvider(project_dir=config_dir, add_global_config=False)]
+        return [string_toml_provider, EnvironProvider(), SecretsTomlProvider(project_dir=config_dir, add_global_config=False), ConfigTomlProvider(project_dir=config_dir, add_global_config=False)]
 
     glob_ctx = ConfigProvidersContext()
     glob_ctx.providers = _initial_providers()
